update game four in line
