+++ src/app/containers/HomeContainer.tsx
@@ -1,8 +1,13 @@
+import { useHomeContainer } from '@core/hooks';
 import { Home } from '@ui/components/Home';
 import { LayoutHome } from '@ui/layouts/LayoutHome/LayoutHome';
 
-export const HomeContainer = () => (
-    <LayoutHome>
-        <Home />
-    </LayoutHome>
-);
+export const HomeContainer = () => {
+    const { onSubmit } = useHomeContainer();
+
+    return (
+        <LayoutHome>
+            <Home onSubmit={onSubmit} />
+        </LayoutHome>
+    );
+};
