+++ src/app/core/hooks/form/index.ts
@@ -1,0 +1,1 @@
+export * from './useForm';
