+++ src/app/core/hooks/home-container/index.ts
@@ -1,0 +1,1 @@
+export * from './useHomeContainer';
