--- src/app/core/hooks/home/index.ts
@@ -1,1 +1,0 @@
-export * from './useHome';
