+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/app/containers/boardcontainer.tsx","./src/app/containers/homecontainer.tsx","./src/app/core/constants/index.ts","./src/app/core/hooks/index.ts","./src/app/core/hooks/board-game/index.ts","./src/app/core/hooks/board-game/useboardgame.tsx","./src/app/core/hooks/home/index.ts","./src/app/core/hooks/home/usehome.tsx","./src/app/core/interfaces/index.ts","./src/app/core/interfaces/state.interface.ts","./src/app/core/interfaces/game/boardgame.interface.ts","./src/app/core/interfaces/game/index.ts","./src/app/core/interfaces/game/player.interface.ts","./src/app/core/interfaces/home/index.ts","./src/app/core/interfaces/home/playerform.interface.ts","./src/app/core/interfaces/layoutboard/index.ts","./src/app/core/interfaces/layoutboard/layoutboard.interface.ts","./src/app/core/interfaces/routes/guardprops.interface.ts","./src/app/core/interfaces/routes/index.ts","./src/app/core/state/appcontext.tsx","./src/app/core/state/reducer.ts","./src/app/core/state/board-game/actions.ts","./src/app/core/state/board-game/index.ts","./src/app/core/types/index.ts","./src/app/core/types/optionscolors.type.ts","./src/app/core/types/square.type.ts","./src/app/core/utils/checkfourinline.ts","./src/app/core/utils/checktie.ts","./src/app/core/utils/checkwinner.ts","./src/app/core/utils/index.ts","./src/app/routes/indext.tsx","./src/app/routes/guards/guardsession.tsx","./src/app/ui/components/board/index.tsx","./src/app/ui/components/game/index.tsx","./src/app/ui/components/headerboard/index.tsx","./src/app/ui/components/home/index.tsx","./src/app/ui/components/resetbutton/index.tsx","./src/app/ui/components/square/index.tsx","./src/app/ui/layouts/layoutboard/layoutboard.tsx","./src/app/ui/layouts/layouthome/layouthome.tsx"],"version":"5.6.3"}
+{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/app/containers/boardcontainer.tsx","./src/app/containers/homecontainer.tsx","./src/app/core/constants/index.ts","./src/app/core/hooks/index.ts","./src/app/core/hooks/board-game/index.ts","./src/app/core/hooks/board-game/useboardgame.tsx","./src/app/core/hooks/form/index.ts","./src/app/core/hooks/form/useform.tsx","./src/app/core/hooks/home-container/index.ts","./src/app/core/hooks/home-container/usehomecontainer.tsx","./src/app/core/interfaces/index.ts","./src/app/core/interfaces/state.interface.ts","./src/app/core/interfaces/game/boardgame.interface.ts","./src/app/core/interfaces/game/index.ts","./src/app/core/interfaces/game/player.interface.ts","./src/app/core/interfaces/home/index.ts","./src/app/core/interfaces/home/playerform.interface.ts","./src/app/core/interfaces/layoutboard/index.ts","./src/app/core/interfaces/layoutboard/layoutboard.interface.ts","./src/app/core/interfaces/routes/guardprops.interface.ts","./src/app/core/interfaces/routes/index.ts","./src/app/core/state/appcontext.tsx","./src/app/core/state/reducer.ts","./src/app/core/state/board-game/actions.ts","./src/app/core/state/board-game/index.ts","./src/app/core/types/index.ts","./src/app/core/types/optionscolors.type.ts","./src/app/core/types/square.type.ts","./src/app/core/utils/checkfourinline.ts","./src/app/core/utils/checktie.ts","./src/app/core/utils/checkwinner.ts","./src/app/core/utils/index.ts","./src/app/routes/indext.tsx","./src/app/routes/guards/guardsession.tsx","./src/app/ui/components/board/index.tsx","./src/app/ui/components/form/index.tsx","./src/app/ui/components/game/index.tsx","./src/app/ui/components/headerboard/index.tsx","./src/app/ui/components/home/index.tsx","./src/app/ui/components/resetbutton/index.tsx","./src/app/ui/components/square/index.tsx","./src/app/ui/layouts/layoutboard/layoutboard.tsx","./src/app/ui/layouts/layouthome/layouthome.tsx"],"version":"5.6.3"}
